Connected TEMP_DATA to P13
On v10, TEMP_DATA was connected to both P12 and P13 (for the one-wire interface to the temp sensor).

v11 only had TEMP_DATA connected to P12.

With this commit (v12), TEMP_DATA is connected to both P12 and P13 (to match the original v10).

Also incremented version to v12 in BRD and SCH.